[#38] refactor: folder directory 구조 수정
